chore: Refactor `app.component.spec.ts` to use `ComponentFixture`
Refactor the app.component.spec.ts file to import ComponentFixture from
the @angular/core/testing module. This change allows for better testing
of the AppComponent. The previous implementation was using
TestBed.createComponent directly, which is not recommended. By using
ComponentFixture, we can access the component instance and perform more
comprehensive tests. This change improves the testability and
maintainability of the code.